Allow specifying -w and -h params to set width/height for edge-based gestures
If -w and -h are set, these values override automatic screen geometry
detection via X/Wayland.  This can be used for two purposes: 1)
In enviroments where screen resolution is not available such as the
framebuffer the size can be passed manually this way and 2) For running
lisgd edge-based gestures on only part of the screen and for debugging.

Signed-off-by: Maarten van Gompel <proycon@anaproy.nl>