Makefile: Refactor variables for more flexibility
The Makefile is currently conflating CFLAGS and LDFLAGS with CPPFLAGS
and LIBS, respectively. This has several disadvantages:

    - The user cannot easily override LDFLAGS without causing linker
      errors as it currently tracks the libraries we link against.

    - The user cannot override the `-g` optimization level passed during
      linking as it is not tracked via LDFLAGS.

    - It is hard to control include paths as these arent tracked via
      CPPFLAGS. As a consequence, compiling support for X11 and/or
      Wayland conditionally is hard.

Refactor the Makefile to fix those problems by introducing CPPFLAGS
and LIBS variables and make better use of CPPFLAGS and LDFLAGS. Like
this the user can easily override both CFLAGS and LDFLAGS to override
defaults without causing the build to break. And second this sets the
stage for optional build-time support of both X11 and Wayland.

Signed-off-by: Patrick Steinhardt <ps@pks.im>
Signed-off-by: Anjandev Momi <anjan@momi.ca>