lisgd: Allow compiling without support for X11
We currently always links against libX11. All this is used for though is
to auto-detect the display resolution in an X11 environment. Ultimately,
this is unused functionality in setups that use Wayland exclusively.

Add a build-time option `WITHOUT_X11`: if defined by the user, then we
disable this code and thus don't need to link against libX11 anymore.
The default is still to compile with X11-support enabled.

Signed-off-by: Patrick Steinhardt <ps@pks.im>
Signed-off-by: Anjandev Momi <anjan@momi.ca>