lisgd: Allow compiling without support for Wayland
Similar to the preceding commit, this commit adds a new build-time
option `WITHOUT_WAYLAND`. If defined by the user, then we disable the
code to auto-detect the resolution in a Wayland environment and thus
don't need to link against libwayland-client anymore.

With this change we can now build lisgd without support for either X11
or Wayland. In this case the user can still manually specify the desired
display resolution.

Signed-off-by: Patrick Steinhardt <ps@pks.im>
Signed-off-by: Anjandev Momi <anjan@momi.ca>